// src/api-client/models/promotion-update-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface PromotionUpdateRequest
 */
export interface PromotionUpdateRequest {
    /**
     * 
     * @type {string}
     * @memberof PromotionUpdateRequest
     */
    'code': string;
    /**
     * 
     * @type {string}
     * @memberof PromotionUpdateRequest
     */
    'description'?: string;
    /**
     * 
     * @type {number}
     * @memberof PromotionUpdateRequest
     */
    'discountPercent'?: number;
    /**
     * 
     * @type {string}
     * @memberof PromotionUpdateRequest
     */
    'startDate'?: string;
    /**
     * 
     * @type {string}
     * @memberof PromotionUpdateRequest
     */
    'endDate'?: string;
}